functionary/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
attrs==22.2.0
    # via jsonschema
billiard==3.6.4.0
    # via celery
celery==5.2.7
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2022.12.7
    # via
    #   minio
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==39.0.1
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.1.6
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-celery-beat
    #   django-htmx
    #   django-picklefield
    #   django-tables2
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-allauth==0.52.0
    # via -r requirements.in
django-bootstrap5==22.2
    # via -r requirements.in
django-browser-reload==1.6.0
    # via -r requirements.in
django-celery-beat==2.4.0
    # via -r requirements.in
django-constance[database]==2.9.1
    # via -r requirements.in
django-htmx==1.13.0
    # via -r requirements.in
django-tables2==2.5.1
    # via -r requirements.in
django-picklefield==3.1
    # via django-constance
django-timezone-field==5.0
    # via django-celery-beat
django-widget-tweaks==1.4.12
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-spectacular
docker==6.0.1
    # via -r requirements.in
drf-spectacular[sidecar]==0.25.1
    # via -r requirements.in
drf-spectacular-sidecar==2022.12.1
    # via drf-spectacular
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.17.3
    # via
    #   -r requirements.in
    #   drf-spectacular
kombu==5.2.4
    # via celery
minio==7.1.13
    # via -r requirements.in
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.0
    # via docker
pika==1.3.1
    # via -r requirements.in
prompt-toolkit==3.0.36
    # via click-repl
psycopg2==2.9.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic==1.10.4
    # via -r requirements.in
pyjwt[crypto]==2.6.0
    # via django-allauth
pyrsistent==0.19.3
    # via jsonschema
python-crontab==2.7.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via python-crontab
python3-openid==3.2.0
    # via django-allauth
pytz==2022.7.1
    # via
    #   celery
    #   django-timezone-field
    #   djangorestframework
pyyaml==6.0
    # via
    #   -r requirements.in
    #   drf-spectacular
rapidfuzz==2.13.7
    # via -r requirements.in
redis==4.5.1
    # via -r requirements.in
requests==2.28.2
    # via
    #   django-allauth
    #   docker
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
sqlparse==0.4.3
    # via django
typing-extensions==4.4.0
    # via pydantic
tzdata==2022.7
    # via django-celery-beat
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.14
    # via
    #   docker
    #   minio
    #   requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
websocket-client==1.5.1
    # via docker
